Rename: 파일 이름 변경
- ProviderChips -> SubscriptionEditor
