+++ src/components/myPage/CurrentSubscriptionInfo.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import styled from '@emotion/styled';
 import { Title, Text, Accordion, Box, Container } from '@mantine/core';
-import { ProviderBadges, SubscriptionProviders, ProviderChips } from './index';
+import { ProviderBadges, SubscriptionProviders, SubscriptionEditor } from './index';
 
 const StyledContainer = styled(Container)`
   background-color: ${({ theme }) => (theme.colorScheme === 'dark' ? theme.colors.dark[6] : theme.colors.gray[1])};
@@ -33,7 +33,11 @@ const CurrentSubscriptionInfo = () => {
             <Text size="2rem">₩29,800</Text>
           </Accordion.Control>
           <Accordion.Panel>
-            {editMode ? <ProviderChips onClick={toggleEditMode} /> : <SubscriptionProviders onClick={toggleEditMode} />}
+            {editMode ? (
+              <SubscriptionEditor onClick={toggleEditMode} />
+            ) : (
+              <SubscriptionProviders onClick={toggleEditMode} />
+            )}
           </Accordion.Panel>
         </Accordion.Item>
       </PresentSubscriptionFee>
