+++ src/components/myPage/index.js
@@ -6,5 +6,5 @@ export { default as ProviderBadges } from './ProviderBadges';
 export { default as Statistics } from './Statistics';
 export { default as SubscriptionProviders } from './SubscriptionProviders';
 export { default as CurrentSubscriptionInfo } from './CurrentSubscriptionInfo';
-export { default as ProviderChips } from './ProviderChips';
+export { default as SubscriptionEditor } from './SubscriptionEditor';
 export { default as PredictedSubscription } from './PredictedSubscription';
